Client-side file distribution according to a neat and organized structure
